feat(record): refine show more/less for large cell content
- drive clip threshold from the visible_cell_height annotation per column
- disable show more/less when the annotation is unset or invalid
- replace the show/collapse button with a hover rail in the entity-key cell
- add chevron caps at both ends of the rail to signal expand/collapse
- only scroll back to the row top on collapse when it scrolled out of view
- flash row-focus only when a collapse actually scrolls
- fix test cases
