+++ test/e2e/locators/record.ts
@@ -92,13 +92,15 @@ export default class RecordLocators {
 
   static getAllColumnNames(container: Locator | Page): Locator {
     return container.locator(
-      'tr:not(.forced-hidden) td.entity-key > span.column-displayname > span'
+      'tr:not(.forced-hidden) td.entity-key .record-entity-key-inner > span.column-displayname > span'
     );
   }
 
   static getColumnNameElement(container: Locator | Page, columnDisplayName: string): Locator {
     const displayName = makeSafeIdAttr(columnDisplayName);
-    return container.locator(`.entity-row-${displayName} td.entity-key > span.column-displayname`);
+    return container.locator(
+      `.entity-row-${displayName} td.entity-key .record-entity-key-inner > span.column-displayname`
+    );
   }
 
   static getAllColumnValues(container: Locator | Page): Locator {
